src/api/src/requirements.txt
anyio==3.6.2
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cryptography==40.0.2
fastapi==0.95.1
google-auth==2.17.3
h11==0.14.0
httptools==0.5.0
idna==3.4
isodate==0.6.1
jsonschema==4.17.3
kubernetes==26.1.0
lazy-object-proxy==1.9.0
oauthlib==3.2.2
openapi-core==0.12.0
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
paramiko==3.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strict-rfc3339==0.7
tapipy==1.2.19
tapisservice==1.2.6
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websocket-client==1.5.1
websockets==11.0.2
